backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
aiohttp==3.10.5
pydantic==2.8.2
python-dotenv==1.0.1
openai>=1.0.0
